# swagger/swagger.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/estado-creacion": {
            "post": {
                "operationId": "EstadoCreacionController_post",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EstadoCreacionDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "estado-creacion"
                ]
            },
            "get": {
                "operationId": "EstadoCreacionController_getAll",
                "parameters": [
                    {
                        "name": "query",
                        "required": false,
                        "in": "query",
                        "description": "query - Filter. e.g. col1:v1,col2:v2 …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fields",
                        "required": false,
                        "in": "query",
                        "description": "fields - Fields returned. e.g. col1,col2 …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sortby",
                        "required": false,
                        "in": "query",
                        "description": "sortby - Sorted-by fields. e.g. col1,col2 …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order",
                        "required": false,
                        "in": "query",
                        "description": "order - Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "required": false,
                        "in": "query",
                        "description": "limit - Limit the size of result set. Must be an integer",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "offset",
                        "required": false,
                        "in": "query",
                        "description": "offset - Start position of result set. Must be an integer",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "populate",
                        "required": false,
                        "in": "query",
                        "description": "populate - show subqueries. Must be a boolean",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "estado-creacion"
                ]
            }
        },
        "/estado-creacion/{_id}": {
            "get": {
                "operationId": "EstadoCreacionController_getById",
                "parameters": [
                    {
                        "name": "_id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "estado-creacion"
                ]
            },
            "put": {
                "operationId": "EstadoCreacionController_put",
                "parameters": [
                    {
                        "name": "_id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EstadoCreacionDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "estado-creacion"
                ]
            },
            "delete": {
                "operationId": "EstadoCreacionController_delete",
                "parameters": [
                    {
                        "name": "_id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "estado-creacion"
                ]
            }
        },
...
        "/horario": {
            "post": {
                "operationId": "HorarioController_post",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/HorarioDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "horario"
                ]
            },
            "get": {
                "operationId": "HorarioController_getAll",
                "parameters": [
                    {
                        "name": "query",
                        "required": false,
                        "in": "query",
                        "description": "query - Filter. e.g. col1:v1,col2:v2 …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fields",
                        "required": false,
                        "in": "query",
                        "description": "fields - Fields returned. e.g. col1,col2 …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sortby",
                        "required": false,
                        "in": "query",
                        "description": "sortby - Sorted-by fields. e.g. col1,col2 …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order",
                        "required": false,
                        "in": "query",
                        "description": "order - Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "required": false,
                        "in": "query",
                        "description": "limit - Limit the size of result set. Must be an integer",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "offset",
                        "required": false,
                        "in": "query",
                        "description": "offset - Start position of result set. Must be an integer",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "populate",
                        "required": false,
                        "in": "query",
                        "description": "populate - show subqueries. Must be a boolean",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "horario"
                ]
            }
        },
        "/horario/{_id}": {
            "get": {
                "operationId": "HorarioController_getById",
                "parameters": [
                    {
                        "name": "_id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "horario"
                ]
            },
            "put": {
                "operationId": "HorarioController_put",
                "parameters": [
                    {
                        "name": "_id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/HorarioDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "horario"
                ]
            },
            "delete": {
                "operationId": "HorarioController_delete",
                "parameters": [
                    {
                        "name": "_id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "horario"
                ]
            }
        }
    },
    "info": {
        "title": "Horarios_crud",
        "description": "API CRUD para la gestion de horarios",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "schemas": {
...
            "EstadoCreacionDto": {
                "type": "object",
                "properties": {
                    "Nombre": {
                        "type": "string"
                    },
                    "Descripcion": {
                        "type": "string"
                    },
                    "CodigoAbreviacion": {
                        "type": "string"
                    },
                    "Activo": {
                        "type": "boolean"
                    },
                    "FechaCreacion": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "FechaModificacion": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": [
                    "Nombre",
                    "Descripcion",
                    "CodigoAbreviacion",
                    "Activo",
                    "FechaCreacion",
                    "FechaModificacion"
                ]
            },
...
            "HorarioDto": {
                "type": "object",
                "properties": {
                    "Nombre": {
                        "type": "string"
                    },
                    "CodigoAbreviacion": {
                        "type": "string"
                    },
                    "Codigo": {
                        "type": "string"
                    },
                    "ProyectoAcademicoId": {
                        "type": "number"
                    },
                    "PlanEstudioId": {
                        "type": "number"
                    },
                    "Semestres": {
                        "type": "number"
                    },
                    "PeriodoId": {
                        "type": "number"
                    },
                    "EstadoCreacionId": {
                        "$ref": "#/components/schemas/EstadoCreacionDto"
                    },
                    "Observacion": {
                        "type": "string"
                    },
                    "Activo": {
                        "type": "boolean"
                    },
                    "FechaCreacion": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "FechaModificacion": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": [
                    "Nombre",
                    "CodigoAbreviacion",
                    "Codigo",
                    "ProyectoAcademicoId",
                    "PlanEstudioId",
                    "Semestres",
                    "PeriodoId",
                    "EstadoCreacionId",
                    "Observacion",
                    "Activo",
                    "FechaCreacion",
                    "FechaModificacion"
                ]
            },
...
        }
    }
}